requirements.txt
certifi==2020.12.5
asgiref==3.3.1
Django==3.1.7
django-environ==0.4.5
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.2
requests==2.25.1
Pillow==8.1.0
